refactor: class 이름 파일 이름 통일
