+++ src/pages/games/SnackGame/game/model/snackGame/SnackGameDefualt.ts
